+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Upgraded
+- Keyple Plugin API: `2.0.0` -> `2.1.0`
+### Upgraded (example app)
+- Calypsonet Terminal Calypso API: `1.6.0` -> `1.8.0`
+- Keyple Service Lib: `2.1.3` -> `2.2.1`
+- Keyple Service Resource Lib: `2.0.2` -> `2.1.1`
+- Keyple Card Calypso Lib: `2.3.2` -> `2.3.5`
+- Keyple Plugin PC/SC Lib: `2.1.0` -> `2.1.2`
 
 ## [1.0.0] - 2023-03-03
 This is the initial release.